+++ src/lib/tabs/navigation.tsx
@@ -1,26 +1,8 @@
 import React from "react";
 import { Route, Routes, useLocation } from "react-router-dom";
 
-import * as T from "../../lib/tabs/type";
-
-const isSelected = (path: string, pathname: string) => pathname.includes(path);
-
-const isSelectedFromArray = (
-  pathname: string,
-  paths: T.TabNavigationProps[],
-  tabPrefix: string
-): string => {
-  // console.log({ paths, pathname, tabPrefix });
-  const f = paths.filter((x) =>
-    isSelected(tabPrefix + (x.path || ""), pathname)
-  );
-
-  if (f.length) {
-    return f[0].path || "";
-  }
-
-  return "";
-};
+import * as T from "../tabs/type";
+import { isSelectedFromArray } from "./utils";
 
 const Navigation =
   (
